Backend_homify/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-filter==23.3
psycopg[binary]>=3.1.18
Pillow
python-decouple==3.8
redis==5.0.1
celery==5.3.4
drf-yasg==1.21.7
